Add ENG major
working, but not tested yet! There is a bug where the req path for english major replaces the ENG in all req paths with Eng so find satisfying courses does not work. With other majors, the major code in the req path remains in all caps so something weird is going on with english.