Remove hardcoded year from req file names
The year in the filenames is no longer used, and so this pull request 
removes it.

The pull request has no real effect, and is just clean-up, but it is 
separated from the other related changes because it could cause conflics 
with other open pull requests which must be resolved.
It can stand alone, and can therefore be merged separately from the 
other PRs in the stack.

Specifically, it renames the files which the remaining open requirement 
update pull requests are modifying.